+++ index.php
@@ -105,5 +105,22 @@ if (!$page) {
 </div>
 <script src="/js/lightbox.js"></script>
 
+<!-- Default Statcounter code for SonaCMS
+https://www.sonacms.com -->
+<script type="text/javascript">
+    var sc_project=13335240;
+    var sc_invisible=1;
+    var sc_security="d57d5ff3";
+</script>
+<script type="text/javascript"
+        src="https://www.statcounter.com/counter/counter.js"
+        async></script>
+<noscript><div class="statcounter"><a title="Web Analytics"
+                                      href="https://statcounter.com/" target="_blank"><img
+                    class="statcounter"
+                    src="https://c.statcounter.com/13335240/0/d57d5ff3/1/"
+                    alt="Web Analytics"
+                    referrerPolicy="no-referrer-when-downgrade"></a></div></noscript>
+<!-- End of Statcounter Code -->
 </body>
 </html>